ELk7: add curator job to require disktype hdd after 7 days
This adds a curator job to require disktype=hdd for indices that are older
than 7 days, and had an existing allocation requirement for disktype "ssd"

The preexisting allocation requirement is intended to work around differences
in the v5 and v7 clusters. More specificaly, in the v5 cluster the routing
allocation require is not set, so this action should be a noop there.

Also switching the order of the jobs because the forcemerge job can be
long-running, and is set to fail upon error.

Bug: T247376
Change-Id: I16372c7b911ae9c6baf8494085e35de069d2e061