mediawiki::jobrunner: raise max_execution_time
While max_execution_time usually underestimates grossly the time spent
in executing a php script (because it doesn't count anything executed
as a C function or a syscall), sometimes jobs need to process php
instructions for longer than our default timeout. So raise it to the
timeout for jobs - no need to get to the same value as videoscaling
timeouts as those run as forked processes, so max_execution_time
doesn't count much.

Bug: T247622
Change-Id: Ic3620f2f52748ebfebdc427272f064e84491e707