Integration Cluster: update gitcache nightly
This saves a fair bit of work on the CI workers as well as on Gerrit by
using "git clone --reference" each time we pull down one of these
repositories we re-use the objects on disk to determine which objects to
send in the response packfile. This saves network traffic and IO on
Gerrit.

Change-Id: Idc96dd6c746223dcf80ee08c6975e5abac0d547b