nova policy.json: replace 'admin_or_member' with 'admin_or_owner'
I don't like the way admin_or_owner is calculated but it's what
all upstream policy files use so we should use it too.  Ultimately
this change should alter any behavior.

This patch also explicitly defines the 'admin_api' rule (which
previously was used but not defined), and removes the 'default'
rule which is no longer used.

Bug: T247573
Change-Id: I8d2e617fdefdb233a76e690d3736337d297de5d2