nova policy.json: Remove all redundant policies
Since Newton, nova has implemented default policies outside of the
policy config.  This patch removes everything from our policy.json
that is a restatement of the same policy in the nova code.

What remains are two categories:

- Things we want to open up to everyone ("")
- Things we want to restrict to only project admins ("rule:admin_or_projectadmin")

This is a big change and very difficult to audit.  I got here by dumping
the in-code policy with oslopolicy-sample-generator and comparing
that file with our file using comm.

Bug: T247573
Change-Id: I2916a9a41d3f5380a936ab0a3e5dcdd588c8a5d7