mariadb-backups: Change bacula backup frequency of dbprov to weekly
Followup to  9dd91cc.

Due to a mistake, daily backups of dbprov snapshots were set. This
is not intended, as only one a week was intended to be stored.

The meaning of frequencies is as follows:
* Hourly: Full weekly, incremental hourly
* Weekly: Only fulls, weekly
* Monthly: Fulls monthly, diffs every other fortnite, incr. daily

Monthly was wrongly chosen instead of Weekly frequency. Correct
so we only perform a backup once a week (it impacts disk spaces
in addition to performance).

Bug: T138562
Change-Id: I1b2b629debfe0174fcd889b9a1a59a0f17326bda