Add initial Prometheus monitoring config for Presto
What:
Add basic prometheus jmx agent monitoring for Presto and apply
in Hadoop test.

Why:
Presto is gaining a lot of traction in Analytics, and it seems
that the prometheus jmx agent should be enough to export meaningful
metrics. This change adds supports for it to Presto and enables it
in Hadoop test.

Bug: T247884
Change-Id: I77927d6cda3290a03bedf09f73cad2b05fae762b