site: add CI ganeti nodes
These are new ganeti nodes for CI.

They were not in site.pp yet but now all nodes
are required to have roles.

This also made them pop-up in the Icinga check

"Ensure hosts are not performing a change
on every puppet run".

Adding the "insetup" role will fix that
and add base::firewalling.

Bug: T228926
Change-Id: Ife767433d9697453e9e54904f09f9210c92e88bd