services_proxy: lower idle timeout below the upstream value
While errors we see now are all upstream faliures rather than
connection failures on the part of the services proxy, it's prudent to
lower the idle timeout below what upstream expects.

We might need to re-add retries as some of the errors seem to be
genuine difficulties of the backend under load in eqiad.

Change-Id: Ie3442aa1ce04e4442220dee013f2e8d0a9520820