Remove apache-fast-test, now replaced by httpbb.
For a limited time the script is replaced with a warning message, so
that anyone still trying to use it will know what's going on; eventually
we'll clean that up too.

The tests have already been ported over to httpbb and can be found under
modules/profile/files/httpbb.

This commit also deletes modules/apache/files/apache-fast-test, which
contained almost-but-not-quite the same code and wasn't used anywhere.

Change-Id: I06b22b5b68b390874c792e7a0c6b496858bcc863