realm.pp: trusted facts unavailable when performing a lookup or pcc
The trusted facts are not available when preforming
`puppet master --compile` against a clean puppetdb or
`puppet lookup --compile`

This patch adds a new variable $_trusted_certname which is populated
with facts['fqdn'] if the trusted facts are not available.

It also adds the bool variable $_trusted which is set based on whether
`$_trusted_certname` is set via the $trusted or $facts hash.

Ref: PUP-5441, PUP-9207

Bug: T248169
Change-Id: Idcd4a387e11200d89aa88a890aa6b5b8485555e7