systemd::timer::job: add a parameter to change the systemd slice
What:
Add a paramter to change the default system.slice used for service
units.

Why:
On the Analytics client nodes there are some systemd timers ran by
users that might cause resoure consumption to rise, but they are not
run under the user.slice (that has custom memory/cpu limits).
This change allows to specify the "Slice" setting in the timer's service
unit if needed.

Change-Id: I03b7b9b63ba21cfe807864bfec4865e0f54d63a0