Horizon: split $version into $horizon_version and $openstack_version
Up until now we've been overriding the generic OpenStack version value
selectively for hosts that deploy Horizon.

I'd like for Horizon to be able to apply policy files from whatever
the current version of OpenStack is on e.g. cloudcontrol1001 while
running a different version of Horizon.  To do that we need the Horizon
puppet class to know both versions.

This patch ought to leave every config/policy/whatever as it was
but allow for updating the policy.yaml config in future patches.

Bug: T247575
Change-Id: I947cfd9bfc61ea9cca3e1305612c9c1790acc721