shinken: remove puppet agent checks
Graphite no longer has this data, the diamond collector for this metric has been
deprecated in favor of prometheus. This patch removes the Shinken checks tied to
these metrics.

This data is now located in prometheus, however at this time there is no global
prometheus server scraping exporters in CloudVPS.

Bug: T210993
Change-Id: I50b3e2102d0112b833467428501990016ba9a5a8