mariadb-backups: Skip Saturday snapshot to prevent a retention of 4 copies
On Monday, dbprov1001 has storage issues because for a few minutes, it has
to store 4 copies of backups (Wed, Thu, Fri, Sat) and 2 parallel copies of
s8 and s4 before compression, being dangerously close to 90% disk usage.

For now, skip the Saturday snapshots, which is the least useful, so a
maximum of 3 copies + ongoing are stored on the dbprov hosts, while
weekly backups are already being sent to bacula for longer term.

The retention is kept at 4 days, we are just avoiding 4 consecutive
day snapshots. This may be reverted at a later time with more hardware.

Bug: T138562
Change-Id: I86e0a68ee886f75e62f173e6ba1ca483b33a98c6