Remove postgres spec tests
These were written for the Postgres version in jessie and never updated
for stretch/buster. As we don't use Postgres on jessie for quite a bit
now, they are not useful anymore. They can be remade from scratch in the
future (several things chaned in Debian postgres packaging) and currently
the outdated tests impede unrelated classes to the classes.

Change-Id: I10bd35eabd1e6333c06699a83ebf1532116c1472