profile::analytics::client::limits: add slice settings for cron
What:
Add Slice=user.slice to the cron.service unit via override.

Why:
If my understanding of overrides is right, simply adding the Slice
override for the cron.service system unit should work.
This is needed since all the user crontabs are running, by default,
under the system.slice, that is not limited.

Change-Id: I974c78a371d272a8e2b3962a540d658014eab74c