mediawiki: Fix reqId in php7-fatal-error.php to consider X-Request-Id
Per T247786, there are already requests to MW in prod that follow the
new code path based on $wgAllowExternalReqID, and thus MW ignores
the UNIQUE_ID set by Apache for those requests. This means the entries
in Logstash from MW and from php7-fatal-error.php were not under
the same reqId for those requests.

This commit fixes that by adding support for X-Request-Id here
when it is present, just like MW.

Bug: T247786
Bug: T201409
Change-Id: I5467c7add805b80e62cf65e0b4d0bed63d16aaad