mariadb-backups: Stop replication for s3 when snapshotting
s3 section seem to be overloading when taking logical dumps, probably due
to the extended locking, due to the many amount of objects per instance.

The proper way to fix this is to split s3 into smaller instances. For now,
disable alerts of replication lag for s3 backup source hosts and ignore
replication lag.

Because we will be ignoring it, plus the performance concerns, stop
replication when snapshotting too.

Bug: T138562
Change-Id: I39ac47c45c720dbcacf473f15a970b16702afad0