toolforge: support canonical redirects in urlproxy
As we introduce the new toolforge.org domain, we need a transition
period where folks can opt-in to having requests for the legacy
https://tools.wmflabs.org/$TOOL/... URLs redirected to the new
https://$TOOL.toolforge.org/... URLs. For things running on the 2020
Kubernetes cluster this can easily be handled using an Ingress object
which is fully maintained by `webservice`. For Grid Engine tools we need
to handle the redirection in the dynamicproxy front proxy.

* proxylistener: check for and process a canonical host statement from
  the client. If present, the provided host will be stored in redis
  associated with the tool name for later use from lua in urlproxy.
* urlproxy.lua: Check for a redirect key matching the current tool and
  URL. If found and not set to the current host, set $redirect_to in the
  nginx context to the new host and the current URL fragment.
* urlproxy.conf: If $redirect_to is non-empty after processing via
  urlproxy.lua, send a 301 redirect to the client with the canonical URL
  for the request.

For toolsbeta support, the two new hiera keys should be set to:
 * profile::toolforge::canonical_domain: 'toolsbeta.wmflabs.org'
 * profile::toolforge::canonical_scheme: 'http://

Bug: T234617
Change-Id: Id1106930dff02f2c11a301b8c983aa9572ad66c1