Avoid overriding Hadoop's core files to allow IPv6
What:
Avoid any override to hadoop-config.sh to allow IPv6.

Why:
Since we are moving to a different Hadoop distribution,
we should avoid to override special files like hadoop-config.sh.
In this case we did it to avoid adding java.net.preferIPv4Stack=false
in multiple places, but after recent refactoring it shouldn't
be a big deal of maintainability anymore.
This change will help in avoiding any hiccups when upgrading to Bigtop.

Bug: T244499
Change-Id: Ie7fb01579b600e2fa8dbcf4a580d0b3a2f0425cc