Set maximum failover retry attempts for HDFS in Hadoop Test
What:
Set the property dfs.client.failover.max.attempts to 3 (was 15)
for HDFS in Hadoop test.

Why:
When the HDFS Namenode is not reachable (for example, due to an update)
all the hdfs cli actions in puppet take a ton of time before failing.
The issue is that the hdfs client tries to failover between namenodes
multiple times (default 15) with exponential backoff. Trying 3 times
seems to be reasonable, evaluating this option before applyting to
the Analytics Prod cluster.

Bug: T244499
Change-Id: Iac5c3a0b104919b5fa0f2742500efe501a79e061