profile::tlsproxy::envoy: allow users to override the cluster addr
This PR adds a new parameter `upstream_addr` to Envoyproxy::Tlsconfig.
this allows users to override the default value of $facts['fqdn'] if
the upstream service is listening on a different interface then the
primary e.g. localhost

The CR also adds a check to ensure any remote_addr value is either an
interface on the host in question or one of fdqn or localhost

Hosts: P:tlsproxy::envoy

Change-Id: Iba0aba0e3245068f1b8a7f968721f10b3b914eb2