contint: add acl package for file permissions tweak
I have the use case to expose a ssh-agent socket into a Docker
container. It is launched by Jenkins as jenkins-slave and the socket is
bind mounted into the container. However the UID is preserved and the
file is only readable by the host user.  Thus a process inside the
container running as nobody is unable to read the file.

The acl package installs setfacl which let us grant read/write access to
the UID the process runs as inside the container:

 setfacl -m user:65534:rw "$SSH_AUTH_SOCK"
 docker run \
   --user=nobody \
   -v "SSH_AUTH_SOCK":/ssh-agent.sock \
   -e SSH_AUTH_SOCK=/ssh-agent.sock

Bug: T210271
Change-Id: I367c5ddb7194cc15e369daec68fbb97d34b1fd4f