trusted['certname']: use facts['fqdn'] instead
due to PUP-9207 it seems best to avoid trusted facts unless
they are absolutely required.  The danger of using the fqdn fact is that
this can be controlled by the user if they have root on a machine.

In the case of apero_cas and kerberos this would simply allow a user t
pretend they where the primary server and not the secondary or vice
versa.  however as the user would already have root the machine this
would not increase there level of privileged

Equally it doesn't appear that changing the value in the spark_job module
would cause any damage.

Hosts: idp1001.wikimedia.org,krb1001.eqiad.wmnet

Change-Id: I17cef50fe928f556bf9a0dcd954d57949c97a71c