systemd: Replace the Datetime regex with a call to systemd-analyze.
The calendar format is complex, and our regex only matches a subset of
valid strings. Rather than make it more complex, or write a parser in
Ruby, this shells out to systemd-analyze, which validates the string.

Systemd::Timer::Interval is still a Pattern, both because it's a simpler
grammar and because systemd-analyze doesn't provide a validator for it.

(This also corrects DateTime to Datetime, as it was called elsewhere.)

Change-Id: Ie561aa4cb2504524b742a4de23689feb1ccfa0ae