toolforge: rebuild the docker::builder setup as buster
More than an OS update, this is to build a builder server that is set up
like a current Toolforge worker's docker setup. This will allow builder
images (and therefore buildpacks and the build of the custom controllers).

This should also be more performant and use the storage driver now firmly
recommended on all systems.

Bug: T248703
Change-Id: I18162ad40fdb7c5fe61ce58c99d410bed2e61166