toolforge-k8s: mount /var/lib/docker on appropriate volume
When the worker nodes were built, we didn't take into account that the
instances don't allocate all disk and we should actually set up an lvm
volume to see to that (since we use the overlay2 storage driver which
uses things on the file level on block level as in the docker module).

This should be applied very carefully, one node at a time per the task
attached (even if scripted) because they need to be depooled first.

Bug: T248702
Change-Id: I03838c1ccc1350d334d0b20ce958bd3164ce25cc