mariadb-backups: Move friday backup to Saturday to avoid 3-day old ones
Because we deleted the Saturday backup to avoid having 3 backups in a
row, now on Monday, we can have 3-day old ones (from Friday). Avoid
that by moving the Friday backup to Saturday, so we always have
a backup at least every 2 days.

This means we will have a snapshot of Thursday for Friday only.
However, I think this balances better the fressness and the storage
usage.

Change-Id: I2905abd1b252d282422579b46a6d518266a132d8