toolforge: update the package_builder role so it can be neatly used
Right now, the puppet prefix uses a set of roles and one profile directly.
That's weird, so this will be able to assume things nicely. This also finishes
what was started in Ic71ded4884751560dc2d16e5cadbde16b7fadc23

I never completed or applied that to the builder host.

Change-Id: Id2ad846d757cbca38e310b46fb6823809de8aae5