maps: add parameter to disable tilerator
What:
Add a new hiera parameter (profile::maps::apps::tilerator_enable)
which controls if tilerator should be running.

Why:
During data reimport (or potentially other maintenance operations)
we need to disable tilerator for a long time (days or even weeks).
Keeping this in puppet instead of manually masking the service
makes for a better documentation of the procedure and keeps the
monitoring coherent with the state of the service.

Bug: T249086
Change-Id: Ic75971babf0de9e25735d0d4575e66f9023ac25c