profile::mail::jumpcloud: add new class to manage jumpcloud aliases
This profile installs a script which will query the jump cloud api and
write out a file containing all the users who should be configured for
mail delivery on the MX servers.

When writing the file if an aliases existed but has been removed from
jumpcloud instead of removing the email address entirely we alias it to
a special email address $removed_alias@$managed_domain.  This is
intended as a stop gap to make sure we don't accidentally remove users
however it

Host: O:mail::mx

Bug: T244792
Change-Id: I3dcfc08b365117b209787e3cb9eab8f279aaaf6f