Remove hosts hiera for cloudmetrics1001 and 1002
These files set the OpenStack version to Mitaka.  As
far as I can tell, that setting is a no-op for Buster hosts
which only install the client packages.

I removed some explicit package includes from graphite::wmcs::archiver
because those packages are already included via
profile::openstack::eqiad1::clientpackages.  There was no conflict
before because it happened that the buster/mitaka version of that
package was a no-op.

Bug: T249058
Change-Id: Ie9ff927524a68a432e37249654f24975edcb752a