Openstack client packages:  define for queens/jessie
This will actually just install whatever packages are available
for Jessie; the clients are forward-compatible with queens APIs,
and installing standard upstream packages seems to be what we
mostly do for client packages.

We need this to nurse along our two Jessie servers (labstore1004
and labstore1005) for just a little bit longer.

Bug: T249058
Change-Id: I8cdbdec47b586f44792c071c3a160d82ea77e8c5