apache: restore redirect from stats.wikipedia.org
At some point in the refactors and migrations in puppet this
redirect stopped working. In 2012 when the code was puppetised
(75ceb6640) it worked, and in 2016 it didn't anymore. There was
still Apache config for handling the VirtualHost (0f5815e9b6),
but betewen 2012 and 2016 some change must've caused another
VirtualHost to match instead as the config became unreachable
and started to be handled instead the "Wiki not found" response
via incubator.wikimedia.org (which is what we see today without
this patch).

Bug: T126281
Change-Id: Ie442ce73c26267e2317f64cb63dd5a77f95f4509