wmfkeystonehooks: encode member names as utf8
Even though it breaks things to have the project names in utf8,
it also breaks things to have the member names as python strings.

"The LDAP protocol states that some fields (distinguished names,
relative distinguished names, attribute names, queries) be encoded
in UTF-8. In python-ldap, these are represented as text (str on Python 3,
unicode on Python 2).
[...]
Thus, attribute values are always treated as bytes. Encoding/decoding to other
formats – text, images, etc. – is left to the caller."

This means that we can and should use plain strings for specifying
labels and when making a DN or RDN for a search. We must also encode all
values we set into the LDAP object as bytes. Think of it as being a bit
like escaping user input before using it in an SQL query or as templated
HTML output.

Bug: T249494
Change-Id: Ia59107961d7d15e77e714464d90f43541def7607