backups: Assume backups have their ssds on sda and sdb for partman
Because sometimes ssds and disk shelves can switch positions,
depending on how they are connected, configured and detected,
those sometimes can be switched for the virtual raid disk, ending
up on random sdX drives.

Asume the ssds are the first 2 drives and use the for root. RAID/
shelves will be configured later, after the system is setup.

Bug: T248934
Change-Id: I6329c25ffd3be74f8fd59ec421a88d9111bb60a9