phabricator: add envoy TLS terminator for aphlict
Configure a second envoy TLS terminator on the phabricator host
in addition to the existing one that does TLS termination
for Phabricator itself, for the aphlict notification service.

Use a different upstream port from 443 which is already in use.

Use the websockets parameter to allow websocket upgrades.

The point is to avoid that aphlict/websocket traffic goes via httpd
to keep it separate from the main Phabricator.

follow-up will be Iabe36e22a9006.

Also, turn all ports into parameters, one for node, one
for envoy and one for admin access.

And fix/add the firewall rules accordingly.

Bug: T238593
Change-Id: I67dfb77b57c4d3e29bf15b24f30569981a9b1eb4