role::elasticsearch::relforge: use kibana-oss instead of kibana
What:
Install the kibana-oss package rather than the kibana package in
the Relforge cluster.

Why:
From ES6+ there was a split in package naming to take into account
Apache2-licensed code (*-oss) from the other one. This change applies
the same setting that it is currently used for ES7. A follow up
change should be needed to move the parameter to a proper hiera
variable.

Bug: T246961
Change-Id: Ic2eef8e5839e75714f80729c4b9bd45931ef2575