[mwgrep] only query live indices
While reindexing CirrusSearch creates a new index, once it's done it switches
the index alias to this new index and removes the old one. But when a reindex
fails this new index is not removed.
By forcing mwgrep to only query indices that ends with (content, general or
file) we ensure that it does not query these stale indices.

Bug: T249435
Change-Id: Ida9f8f4948326a7983070360a5ed572511705fd4