profile::kubernetes: add the puppet CA cert to general.yaml
Currently we need to inject the puppet CA certificate into
changeprop's container. This pattern already exists for the eventgate
services in a few places - this means that the certificate appears
repeatedly in the values.yaml files and could need to be added to
others in future. Rather than perpetuate this pattern, make the cert
contents available to all charts by default.

Bug: T249633
Change-Id: Iabf5f6fd6ceab948f3b42eb44202e47c76820d1d