analytics::refinery::eventlogging-saltrotate: Bootstrap salts
Before, this script assumed EventLogging salts were always present
under <%= @refinery_config_dir %>/salts/eventlogging_sanitization.
This is error prone, if the machine goes down or whenever we
want to migrate this service to another host.

This change makes the script pull the salts from HDFS if they are
not already in the local salt directory (bootstrapping).

Change-Id: I36ca458c191951b0c61312a7c06d39a06f229b9a