netbox: switch git::clone to ensure present, not latest
There are 2 reasons that make me suggest to stop using "latest" here.

The first one is T218900 which suggest to reduce or outright
ban the use of it in production.

And the other is that this means that puppet repeats a change
on every single run, making it trigger this Icinga alert:

https://icinga.wikimedia.org/cgi-bin/icinga/extinfo.cgi?type=2&host=puppetdb1002&service=Ensure+hosts+are+not+performing+a+change+on+every+puppet+run

Change-Id: I9b0111b893d16db88d1eb83bedff4dd8186b5e69