zuul: use zuul-deployers not contint-root
The ssh key for Zuul in keyholder is restricted to members of
contint-roots however that group is not on the deployment servers. We
can not add it cause it adds extra sudo rules to those members which are
not suitable for the deployment server.

Create new zuul-deployers group that does not have any sudo rule
attached. A YAML alias ensure that all contint-roots are in it.

Add the group to the depoyment server (via admin::groups).

Modelled after aa29ce7d03993750d28eeae06a55609a1d0a5ecf which was for
Gerrit.

Hosts: deploy1001.eqiad.wmnet

Bug: T224591
Change-Id: I7cbb7e59813b53617c2a32ff5aafd183cc744c80