merge microsites into webserver_misc_apps
Traditionally we had 2 separate pairs of VMs for misc sites,
one for apps and one for static sites.

The suggestion here is to merge them into a single set of VMs,
add all static site profiles to webserver_misc_apps
and then eventually remove VMs bromine and vega which are on stretch.

The miscweb* have recently been upgraded to buster.

Bug: T247650
Change-Id: If1e9229d46e7ddcf5d4314f06ee43d7f2f26ff80