icinga: Add git local changes check
Creates modules/icinga/files/check_git_local_changes.py

This checks a specified path and alerts on untracked files, with optional age
specifier so that untracked files over a certain age will cause an alert, but
otherwise cause a warning. It also checks if there are unstaged or
uncommited local changes.

The use case of this check is for git repositories that may accumulate manual
changes inappropirately during testing.

Change-Id: I417767c8d888ef48dc3b6ec815f1b7a659a1be9c