mediawiki::php: prevent max_execution_time from triggering
We don't really want to use max_execution_time, because it's both a
bad metric (it doesn't count wall clock) and cannot stop execution if
we're in a syscall or a C function. We use a combination of fpm's
request_timeout and of excimer's own timeouts instead. So raise
max_execution_time to an arbitrary higher value than the
request_timeout, so it won't kick in.

Bug: T248564
Change-Id: I908c2fa05f0952bf2d953fd6f802e2300b48b5c4