safe-service-restart: allow for a grace period after depooling
Sometimes we might want to allow for a grace time when restarting a
service, between when we depool it and when we actually restart it,
allowing for a good chunk of incoming requests to be completed.

In this change we only add the cli argument and a conditional sleep,
so it should be a noop for all callers.

Change-Id: I92eb7c94f7522c3ef449537d0c0d0078d3ac8b1b