contint: remove unused Gerrit firewall rule
The CI masters have a rule to allow connections from Gerrit servers to
port 29418.  There is no such service on the CI masters.

e3aaad272e6 added the rule in 2013 for "git replication purposes". We do
not use replication to the CI masters and it is not done over 29418
anyway.

The rule is unused (first two columns are packets and bytes:

 contint1001$ iptables --list -v|grep :29418
   0  0 ACCEPT  tcp  any  any  gerrit1001.wikimedia.org  tcp dpt:29418
   0  0 ACCEPT  tcp  any  any  gerrit2001.wikimedia.org  tcp dpt:29418
   0  0 ACCEPT  tcp  any  any  gerrit.wikimedia.org  tcp dpt:29418

Change-Id: I7d4251d32fcfc4f65c16a889a1954f11fff5ebc9