mariadb-backups: Split mariadb::backups into 2 roles for backup2002
backup2002 is used for content (External Storage) servers. While
the functionality there is almost identical, create a separate role
for it, as:
* They will not generate snapshots
* They will have a slightly different schedule and retention policy
  (to be implemented)
* They will be not only clients, but also storage daemons for bacula
  (to be implemented- for now remove all bacula references)

This will allow to reuse the profiles, but have a different role
for each of them.

Metadata and misc backups will be setup on dbprov servers with the
mariadb::backups role.
Content backups will be setup on bacula[12]002 with the
mariadb::content_backups role.

We can rename mariadb::backups later if we want.

Bug: T79922
Change-Id: I678ee4e3e80720caec54751be48da0a8a25f1bc5