Use -XX:NewRatio=3 for cloudelastic-chi instead of setting a specific size
What:
Use NewRatio (JVM Option) instead of setting a specific size for the young gen
on cloudelastic-chi.

Why:
It seems more convenient to use a ratio rather than a specific value, more
flexible and manageable. Since there was already an option to tune NewRatio,
I refactored it to be more generic and accept more values.

Bug: T231517
Change-Id: I5dc13de9d095ff1087b2cee91620be2a3d2532d1