Refactor logstash::input::kafka to DRY ssl_truststore_location logic
Instead of requiring users of logstash::input::kafka to manually
deploy the target Kafka cluster's truststore, use the
kafka_config function to infer the truststore location
and deploy it into /etc/logstash with a predictable name.

Now, all users of logstash::input::kafka have to do to use SSL is
provide the $kafka_cluster_name and set $security_protocol = 'SSL'.

In a future patch I'm about to add Kafka logstash inputs using SSL
for all Kafka clusters, and it will be easiest not have to manually
deploy the truststores for them when I do.

This is mostly a no-op, but will rename the paths for the existent
truststore.jks files on logstash collector nodes.

Change-Id: I16766c90ea40c07ff7220852a5edf3e7b68e354c