Neutron.conf: change rpc_timeout to 30
The default timeout here is 60.  It was cut to 3s in

  https://gerrit.wikimedia.org/r/#/c/operations/puppet/+/463062/

In Rocky something is slower, and a timeout of 3 causes
the code that stands up a new l3 agent to time out
(in neutron/agent/l3/agent.py in _process_added_router())

Raising this back to 30 gets our agents set up properly.  We're hoping
that the initial problem is reduced by tuning we've done elewhere and
this won't cause the old bug (T205524) to reappear.

Bug: T205524
Bug: T250428
Change-Id: Ic3af2df737ddecd2d1ed0c2715ba15cf42170ed0