elasticsearch::instance: avoid UseCMSInitiatingOccupancyOnly
What:
When using the CMS GC, avoid +UseCMSInitiatingOccupancyOnly.

Why:
The +UseCMSInitiatingOccupancyOnly for CMS is very powerful, because
it forces the GC to use the CMSInitiatingOccupancyFraction value as
only metric to start an old gen collection. In our case this means
that if the old gen pool is filled up more than 75%, an old gen
collection is forced even if might not be needed. For example, in
heavy used and small clusters like cloudelastic, during indexations
we observed that the old gen pool is not used fully, but keeps triggering
old gen collections since the usage is a little bit more than 75%.
Removing +UseCMSInitiatingOccupancyOnly should be ok, the JVM probably
knows better when to trigger an old gen collection.

Bug: T231517
Change-Id: I985531fa98de9e41a35ea4ff5b3b6de432ee9952