mariadb-backups: Skip generation of backups for read-only es sections
After a first run of backups covering all es hosts has been done,
stop genereting more backups of es1, es2, and es3, which are supposed
to be read only, and never changing (they could be enabled again in
the future if they are lost, or modified for any reason).

For now on, take only backups of es4 and es5, the only 2 read-write
replica sets, every week.

Bug: T79922
Change-Id: I2de5b50bcc460cbefef2ff5a77b49e43b15b23a6