tlsproxy::envoy: allow limiting firewall srange
When using this profile on hosts with public IPs
we would currently open the port to the world.

Add an optional parameter to specify an srange
for ferm/iptables.

Default value stays as undef to not affect existing setups.

Limit the allowed source range to caching servers on
contint masters.

Bug: T149804
Bug: T224591
Change-Id: Id105b4b82684b7b1da58e81bf0b15f67f6820070