mediawiki::mcrouter_wancache: add a failover route to a secondary pool
We've had several cases where the unavailability of one memcached
server caused a severe impact on the latency and availability of
MediaWiki.

This patch introduces (for now, only on mwdebug1001) a
failover pool (the so-called "gutter pool") that will be used by
mcrouter whenever one server in the main pool becomes unavailable.

We chose to set an expiry time to the keys in the failover pool of 600
seconds, which should be enough to guarantee availability in an
emergency, and not enough to cause long-lasting consistency troubles.

Bug: T213089
Change-Id: I4e8124fb17cd46c3ab41732c30b702a889b48588