contint: ignore more Docker partitions disk checks
We ignore disk check for /run/docker and /mnt/docker. Add to the ignore
list a couple others:

* /var/lib/docker the default from the Debian package. The actual
partition is / which is already monitored.  If we have the Docker data
on a different disk it would use a different path such as /mnt/docker on
contint1001

* Ignore /srv/docker which is the path that is going to be used on
contint2001.wikimedia.org using the extra disk mounted at /srv.

Hosts: contint1001.wikimedia.org, contint2001.wikimedia.org

Bug: T224591
Change-Id: I96ad15446dfec70bcaca83af518e036ffb1af7d9