contint: move Docker data out of / on contint2001
The Debian package defaults to use /var/lib/docker which on contint2001
is on the / partition.

Move Docker data files to /srv/docker which is on the dedicated disks
mounted at /srv.

That varies compared to contint1001 which has an extra pair of disks on
/mnt.

Hosts: contint1001.wikimedia.org, contint2001.wikimedia.org

Bug: T224591
Change-Id: I832e5d97a19bc1d4da58a2abeb9fce477588bfab