ATS: switch backends for annual.wm.org and 15.wp.org
https://annual.wikimedia.org and https://15.wikipedia.org
were so far hosted on the "misc static" VMs but we are
now merging these static microsites into the "misc apps" VMs
which are already running on buster.

/srv/org/wikimedia/annualreport/ on the old (bromine/vega)
and new (miscweb1002/miscweb2002) are identical since puppet
just git clones the contents.

15.wikipedia.org just points to /srv/org/wikimedia/annualreport/2015
so is part of it with a special URL.

Bug: T247650
Change-Id: I07c3f77efc5b8c8d8620d7ee16787292078cdbe0