toolforge-k8s: add a puppetized checkout of maintain-kubeusers
While the image is built elsewhere, the kustomized yaml is part of deployment
of this core service. This yaml needs to be on the control nodes or another
location with admin deploy functionality, so it has so far been in my user
home directory. It seems more appropriate to check it out on the node.

This will also facilitate moving its needed yaml into kustomize per
I6f8dd2fd5bd861c4aa2f69ec36d. This way, yaml that is used to control
a core service for the cluster is only in a single repo, and it is moved
to the repo of the service that requires it.  On that note, I'd also
like to do the same with the yaml and scripts for the custom controllers.

Change-Id: I05f8cc3529694f90d906edd813520fb40f0b6a8f