gerrit: add Keepalive=on to ProxyPass config lines
Trying to debug the issue described in T246763 where

"the backend server closed the pooled connection after
the connection check by the proxy and before data sent by
the proxy reached the backend"

leading to occasional 502s.

I found:

https://serverfault.com/questions/185894/proxy-error-502-reason-error-reading-from-remote-server-with-apache-2-2-3-de

where a user is in the same situation with Apache proxying to jetty
and gets the same "error reading from remote server".

He describes trying some tricks such as "proxy-initial-not-pooled"
and "keepalive Off" just like as suggested by Tyler in I99714732b3f6e5,
but also says these did not help.

Then he added a comment how he solved the problem by adding Keepalive=On
to the ProxyPass config lines.

httpd docs describe this as "should be used when you have a firewall
between your Apache httpd and the backend server, which tends to drop inactive
connections".

Despite the firewall comment it does make sense to me to have keepalive
set to on and not off because we want to prevent the connection from
being closed early.

Bug: T246763
Change-Id: I904d8769947c459b1e3b6a442a593c50261644c8