mariadb-backups: Include es4/5 into backup checks and refactor
es4 and es5 started being generated last week. However, they were
still not automatically checked on icinga, despite metadata being
available.

Add es4 and es5 into the list of backup checks.

Because es4 and es5 are not yet setup on eqiad (waiting on
T238602), a refactored was required so eqiad and codfw backups
could be configured independently, with different checks.

Bug: T79922
Change-Id: I4bcf1e83584765d28e3ccf840d4de8f228b7c5c3