role::statistics::explorer: move target hosts to hiera
What:
Move target hosts for various profiles to hiera for visibility.

Why:
For historical reasons, the stat100x hosts have been running with
different roles and now we are trying to unify them, to give a single
standard environment to our users. Some specific bits still needs to
be selectively activated for some stat hosts only, so in this transition
phase we are using hostnames to decided if a functionality needs to be
deployed or not. This changes moves all the tunables to hiera for
visibility.

Bug: T243934
Change-Id: I06aa4eb8edc80899d73866d63915f1f23b0c2479