toolforge: wmcs-package-build: don't use the $INSTANCEPROJECT env var
The $INSTANCEPROJECT env var is not available when we do ssh because the bashrc
file is not loaded in time, even if we instruct ssh to use an interactive
session by using the `-t` switch.
Using a harcoded value of 'tools.admin' should be a good default anyway.

While at it, add proper escaping to the hostname -f call so it executes in the
remote shell rather than locally.

Bug: T249837
Change-Id: Ifb9492b1980c6a48d4d1d23547ef85d425b88138
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>