mariadb-backups: Check backups size also based on previous runs
There was already a backup check that alerted of very small
backups.

However, beyond the trivial case (backups of 0 size) it wasn't
super-useful because sections were very different among each
other. That means that, either it was too easy to generate
false positives (small backups such as tendril) or may not
caught a 50% reduction on backup size due to corruption (or
other backup issue).

While a minimum size is kept to continue detecting 0 size
backups, introduce the concept of critical and warning
"percentage" size limits, where a change of a X% in size,
compared to a previous run, will either give a warning or a
critical alert, depending on the configured threshold.

For now, the default has been set on 1% for warnings and 10% for
criticals, to be tuned later on puppet, based on real data.
In the future, it could also be calculated based on the average
size over a period of time, not just the last sucessful run.

As an additional thing, if there is only 1 previous backup, it
will warn (it already showed critical if 0 backups were found).
This is becase if there is only 1 previous backup, we cannot
compare if the size is reasonable or not, so we ping a human
for manual checking.

Bug: T138562
Change-Id: I4b66924d205b930682df6188373c69294f4b140d