Revert "Update .ruby-version to what is running in production"
rubocop looks at the .ruby-version file to find out which ruby version
to target when linting files.

We use rubocop 0.49.1 which does not have support for ruby 2.5 yet. We
would first need rubocop 0.52.0 to bring in:
https://github.com/rubocop-hq/rubocop/pull/5101

When the file got updated, rubocop has not been run by CI since
rake_modules/taskgen.rb does not consider a change to .ruby-version to
be sufficient to run it.  But that breaks any future changes touching a
ruby file.

This reverts commit 037adb57222b3c3f1b13a0590c0185456aefea14.

Bug: T250538
Change-Id: If45d94da22e8e8d898ba6bd9d73a3be8ea244812