create simplelamp2, replace mysql->mariadb, apache->httpd
A modernized version of the old "simplelamp" role used
across cloud VPS that replaces the mysql module with
mariab and the apache module with httpd.

Both old modules we want to get rid of and this should let users
of the old role have something to replace it with and unblock
us from deleting them.

Bug: T162070
Change-Id: I4a9265b43eeb119e9c9cf25d22df2f40ba9520b9